fix error occuring when no items to delete in delete service

+++ src/Services/DeleteService.php
@@ -39,6 +39,8 @@ class DeleteService
                 $ids[] = $diff['id'];
             }
 
+            if (empty($ids)) return;
+
             $query = sprintf(
                 'DELETE FROM %s WHERE id IN (%s)',
                 $tableName,
